+++ learn/src/main/java/org/platanios/learn/classification/LogisticRegressionSGD.java
@@ -147,7 +147,7 @@ public class LogisticRegressionSGD extends AbstractTrainableLogisticRegression {
      * builder classes.
      */
     public static class Builder extends AbstractBuilder<Builder>
-            implements TrainableClassifier.Builder<Vector, Integer> {
+            implements TrainableClassifier.Builder<Vector, Double> {
         /**
          * Constructs a builder object for a binary logistic regression model that will be trained with the provided
          * training data using the stochastic gradient descent algorithm.
